+++ server/src/test/parser.test.ts
@@ -4,7 +4,6 @@ import * as Path from 'path';
 import { IoService } from '../services/ioService';
 import * as assert from 'assert';
 import { SymbolKind } from 'vscode-languageserver';
-import { stderr } from 'process';
 
 suite('AHK regex parser test', () => {
 	const ioService = new IoService();