+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 1. Command parse
 2. Unit test
 3. fix command parse bug (0.5.1)
+4. improve regex parser. Handle `{}` better. (0.5.3)
+5. Experimental hint for where symbol is included on completion (0.5.3)
 
 ## 0.4.0